+++ java/src/hype/util/HConstants.java
@@ -86,5 +86,6 @@ public interface HConstants {
 		// Misc Math constants
 		SQRT2 = 1.4142135623730951f,
 		PHI = 1.618033988749895f,
-		PHI_1 = 0.618033988749895f;
+		PHI_1 = 0.618033988749895f,
+		EPSILON = (float)10e-12;
 }